requirements.txt
fastapi
uvicorn
python-multipart
requests
pdf2image
google-generativeai
pydantic
python-dotenv
Pillow
requests
pydantic
fastapi
pdf2image
google.generativeai
dotenv